+++ apps/api/src/modules/app-log/app-log.service.ts
@@ -31,8 +31,13 @@ export class AppLogService {
     });
   }
 
-  async log(createLogDto: CreateLogDto): Promise<void> {
+  async log(createLogDto: CreateLogDto, user: Express.User): Promise<void> {
     const logMessage = `${createLogDto.level}: ${createLogDto.message}`;
-    this.logger.info(logMessage);
+    this.logger.info({
+      message: logMessage,
+      stack: createLogDto.stack,
+      path: createLogDto.path,
+      userId: user.id,
+    });
   }
 }