+++ apps/app/src/components/content-wrapper.component.tsx
@@ -11,7 +11,7 @@ const ContentWrapper: React.FC<ContentWrapperProps> = ({ children, padTop = true
       {...props}
       style={{
         padding: padTop ? '1rem' : '0 1rem 1rem 1rem',
-        backgroundColor: '#F0F2F5',
+        backgroundColor: '#f5f5f5',
         display: 'flex', // Set the display to flex
         flexDirection: 'column', // Stack children vertically
       }}