+++ apps/app/src/routes/job-profiles/saved-job-profiles.page.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import { useParams, useSearchParams } from 'react-router-dom';
 import LoadingComponent from '../../components/app/common/components/loading.component';
 import { PageHeader } from '../../components/app/page-header.component';
 import ContentWrapper from '../../components/content-wrapper.component';
@@ -7,8 +7,10 @@ import { useGetSavedJobProfileIdsQuery } from '../../redux/services/graphql-api/
 import JobProfiles from './components/job-profiles.component';
 
 export const SavedJobProfilesPage = () => {
+  const { number } = useParams();
   const [searchParams] = useSearchParams();
   const { data: savedJobProfileIds, isLoading: isSavedJobProfileIdsLoading, refetch } = useGetSavedJobProfileIdsQuery();
+  const page_size = import.meta.env.VITE_TEST_ENV === 'true' ? 2 : 10;
 
   useEffect(() => {
     refetch();
@@ -25,7 +27,9 @@ export const SavedJobProfilesPage = () => {
         <JobProfiles
           key={'SavedProfiles'}
           searchParams={searchParams}
+          selectProfileNumber={number}
           loadProfileIds={savedJobProfileIds?.getSavedJobProfileIds}
+          page_size={page_size}
         />
       </ContentWrapper>
     </>